p138476: count 배열 long 타입으로 수정 / 73.5 점 실패는 전부 런타임 에러

+++ src/main/java/org/example/solution/programers/lv2/p138476/Solution_138476.java
@@ -11,7 +11,7 @@ import java.util.*;
 
 public class Solution_138476 {
     public int solution(int k, int[] tangerine) {
-        int[] count = new int[100001];
+        long[] count = new long[100001];
 
         for (int i = 0; i < tangerine.length; i++) {
             int TG = tangerine[i];
@@ -20,14 +20,13 @@ public class Solution_138476 {
         }
         Arrays.sort(count);
 
-        int sum = 0;
+        long sum = 0;
         int answer = 0;
         for (int i = 100000; i >= 0; i--) {
             answer++;
             sum += count[i];
             if (sum >= k) break;
         }
-
         return answer;
     }
 }
